Consolidate aux signals
